requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow-cpu==2.14.0
librosa==0.10.0
numpy==1.24.3
matplotlib==3.8.0
scipy==1.11.0
reportlab==4.0.0
Pillow==10.0.0
gunicorn==21.2.0
